[Ref] Make TlsServer_client cmake project global
Before the name of the cmake lib to be included by a client of the TlsServer was generated, based upon the name given to TlsServer_DeclareCAmkESComponent().
This is not necessary, as the client library is always identical. So the client lib project was defined with a fixed name TlsServer_client